+++ src/frydom/hydrodynamics/seakeeping/linear/excitation/FrLinearHDBForce.cpp
@@ -58,9 +58,9 @@ namespace frydom {
     // --> Getting sizes
 
     auto nbFreqInterp = waveFrequencies.size();
-    auto nbFreqBDD = BEMBody->GetNbFrequencies();
+    auto nbFreqBDD = m_HDB->GetFrequencyDiscretization().size();// GetNbFrequencies();
     auto nbDirInterp = waveDirections.size();
-    auto nbForceMode = BEMBody->GetNbForceMode();
+    auto nbForceDOFs = GetBodyMask().GetNbDOF(); //->GetNbForceMode();
 
     // Wave direction is expressed between 0 and 2*pi.
     for (auto &dir : waveDirections) dir = mathutils::Normalize_0_2PI(dir);
@@ -70,20 +70,23 @@ namespace frydom {
 
     // -> Building interpolator and return vector
 
-    auto freqsBDD = std::make_shared<std::vector<double>>(BEMBody->GetFrequencies());
+//    auto freqsBDD = std::make_shared<std::vector<double>>(BEMBody->GetFrequencies());
+    auto frequencies = m_HDB->GetFrequencyDiscretization();
+    std::vector<double> vec(frequencies.data(), frequencies.data() + frequencies.rows() * frequencies.cols());
+    auto freqsBDD = std::make_shared<std::vector<double>>(vec);
 
     auto freqCoeffs = std::make_shared<std::vector<std::complex<double>>>();
     freqCoeffs->reserve(nbFreqBDD);
 
     for (auto direction: waveDirections) {
 
-      auto excitationForceDir = Eigen::MatrixXcd(nbForceMode, nbFreqInterp);
+      auto excitationForceDir = Eigen::MatrixXcd(nbForceDOFs, nbFreqInterp);
 
-      for (unsigned int imode = 0; imode < nbForceMode; ++imode) {
+      for (unsigned int idof = 0; idof < nbForceDOFs; ++idof) {
 
         freqCoeffs->clear();
         for (unsigned int ifreq = 0; ifreq < nbFreqBDD; ++ifreq) {
-          freqCoeffs->push_back(m_waveDirInterpolators[imode][ifreq](direction));
+          freqCoeffs->push_back(m_waveDirInterpolators[idof][ifreq](direction));
         }
 
         auto freqInterpolator = mathutils::Interp1dLinear<double, std::complex<double>>();
@@ -91,7 +94,7 @@ namespace frydom {
 
         auto freqCoeffsInterp = freqInterpolator(waveFrequencies);
         for (unsigned int ifreq = 0; ifreq < nbFreqInterp; ++ifreq) {
-          excitationForceDir(imode, ifreq) = freqCoeffsInterp[ifreq];
+          excitationForceDir(idof, ifreq) = freqCoeffsInterp[ifreq];
         }
       }
       Fexc.push_back(excitationForceDir);
@@ -106,19 +109,23 @@ namespace frydom {
     // BEMBody.
     auto BEMBody = m_HDB->GetBody(GetBody());
 
-    auto nbWaveDirections = BEMBody->GetNbWaveDirections();
-    auto nbFreq = BEMBody->GetNbFrequencies();
-    auto nbForceModes = BEMBody->GetNbForceMode();
+    auto nbWaveDirections = m_HDB->GetWaveDirectionDiscretization().size(); //BEMBody->GetNbWaveDirections();
+    auto nbFreq = m_HDB->GetFrequencyDiscretization().size(); //BEMBody->GetNbFrequencies();
+    auto nbForceDOFs = GetBodyMask().GetNbDOF(); //BEMBody->GetNbForceMode();
 
     m_waveDirInterpolators.clear();
-    m_waveDirInterpolators.reserve(nbForceModes);
+    m_waveDirInterpolators.reserve(nbForceDOFs);
 
-    auto angles = std::make_shared<std::vector<double>>(BEMBody->GetWaveDirections(mathutils::RAD, NWU));
+//    auto angles = std::make_shared<std::vector<double>>(BEMBody->GetWaveDirections(mathutils::RAD, NWU));
+
+    auto waveDirections = m_HDB->GetWaveDirectionDiscretization();
+    std::vector<double> vec(waveDirections.data(), waveDirections.data() + waveDirections.rows() * waveDirections.cols());
+    auto angles = std::make_shared<std::vector<double>>(vec);
 
     auto interpolators = std::vector<mathutils::Interp1dLinear<double, std::complex<double>>>();
     interpolators.reserve(nbFreq);
 
-    for (unsigned int imode = 0; imode < nbForceModes; ++imode) {
+    for (unsigned int idof = 0; idof < nbForceDOFs; ++idof) {
 
       interpolators.clear();
 
@@ -129,7 +136,8 @@ namespace frydom {
 
         for (unsigned int iangle = 0; iangle < nbWaveDirections; ++iangle) {
           auto data = GetHDBData(iangle);
-          coeffs->push_back(data(imode, ifreq));
+//          coeffs->push_back(data(idof, ifreq));
+          coeffs->push_back(data(GetBodyMask().GetDOFs()[idof].GetIndex(), ifreq));
         }
 
         auto interpolator = mathutils::Interp1dLinear<double, std::complex<double>>();
@@ -156,9 +164,6 @@ namespace frydom {
                                                             eqFrame->GetFrame().GetY(NWU),
                                                             NWU);
 
-    // DOF.
-    auto nbMode = m_HDB->GetBody(body)->GetNbForceMode();
-
     // Number of wave frequencies.
     auto nbFreq = waveField->GetWaveFrequencies(RADS).size();
 
@@ -166,15 +171,6 @@ namespace frydom {
     auto nbWaveDir = waveField->GetWaveDirections(RAD, NWU, GOTO).size();
 
     // Fexc(t) = eta*Fexc(Nemoh).
-    Eigen::VectorXd forceMode(nbMode);
-    forceMode.setZero(); // Initialization.
-    for (unsigned int imode = 0; imode < nbMode; ++imode) {
-      for (unsigned int ifreq = 0; ifreq < nbFreq; ++ifreq) {
-        for (unsigned int idir = 0; idir < nbWaveDir; ++idir) {
-          forceMode(imode) += std::imag(complexElevations[idir][ifreq] * m_Fhdb[idir](imode, ifreq));
-        }
-      }
-    }
 
     // From vector to force and torque structures.
     Force force;
@@ -182,18 +178,24 @@ namespace frydom {
     Torque torque;
     torque.SetNull();
 
-    for (unsigned int imode = 0; imode < nbMode; ++imode) {
-
-      auto mode = m_HDB->GetBody(body)->GetForceMode(imode);
-      Direction direction = mode->GetDirection(); // Unit vector for the force direction.
-      switch (mode->GetType()) {
-        case FrBEMMode::LINEAR:
-          force += direction * forceMode(imode);
+    unsigned int idof = 0;
+    for (auto &dof:GetBodyMask().GetDOFs()) {
+      double tempforce = 0;
+      for (unsigned int ifreq = 0; ifreq < nbFreq; ++ifreq) {
+        for (unsigned int idir = 0; idir < nbWaveDir; ++idir) {
+          tempforce += std::imag(complexElevations[idir][ifreq] * m_Fhdb[idir](idof, ifreq));
+        }
+      }
+      Direction direction = dof.GetDirection();
+      switch (dof.GetType()) {
+        case HDB5_io::DOF::LINEAR:
+          force += direction * tempforce;
           break;
-        case FrBEMMode::ANGULAR:
-          torque += direction * forceMode(imode);
+        case HDB5_io::DOF::ANGULAR:
+          torque += direction * tempforce;
           break;
       }
+      idof++;
     }
 
     // Projection of the loads in the equilibrium frame.
@@ -215,6 +217,10 @@ namespace frydom {
                                      const std::shared_ptr<FrHydroDB> &HDB) :
       FrForce(name, type_name, body),
       m_HDB(HDB) {}
+
+  FrMask FrLinearHDBForce::GetBodyMask() const {
+    return m_HDB->GetBodyDOFMask(m_HDB->GetBody(GetBody()));
+  }
 
 
 } // end namespace frydom