+++ tools/frydom/HDB5tool/HDB5tool.py
@@ -227,8 +227,7 @@ def get_Arg_part_1_CE(args, database):
         database.discretization._delta_time = float(args.time_step_irf[0])
 
     # Initialize pyHDB.
-    if (
-            args.path_to_nemoh_cal is not None or args.initialization is True):  # _initialize is automatically called when a .cal is read.
+    if (args.path_to_nemoh_cal is not None or args.initialization is True):  # _initialize is automatically called when a .cal is read.
         database._initialize()
 
     # Body - Active hydrostatics (useless).